pyproject.toml
[build-system]
requires = ["setuptools >= 64", "setuptools_scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clin34"
dynamic = ["version"]
description = "clin34 benchmarking"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    { name = "Bram Vanroy", email = "bramvanroy@hotmail.com" },
]
maintainers = [
    { name = "Bram Vanroy", email = "bramvanroy@hotmail.com" },
]

keywords = [
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing",
]
dependencies = [
    "accelerate==0.33.0",
    "bitsandbytes==0.43.3",
    "datasets==2.20.0",
    "hf_transfer==0.1.8",
    "numpy==1.26.4",
    "openpyxl==3.1.5",
    "outlines==0.0.46",
    "pandas==2.2.2",
    "protobuf==5.27.3",
    "PyYAML==6.0.2",
    "sacremoses==0.1.1",
    "sentencepiece==0.2.0",
    "scikit-learn==1.5.1",
    "torch==2.4.0",
    "tqdm==4.66.5",
    "transformers==4.44.0",
    "typer==0.12.3"
]

[project.optional-dependencies]
dev = [
    "ruff"
]

[project.urls]
homepage = "https://github.com/BramVanroy/clin34-benchmarks"
documentation = "https://github.com/BramVanroy/clin34-benchmarks"
issues = "https://github.com/BramVanroy/clin34-benchmarks/issues"
repository = "https://github.com/BramVanroy/clin34-benchmarks.git"

# TOOLS
[tool.setuptools]
include-package-data = true
license-files = ['LICEN[CS]E*']

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_file = "src/clin34/version.py"

[tool.ruff]
line-length = 119

[tool.ruff.lint]
# Never enforce `E501` (line length violations).
ignore = ["C901", "E501", "E741", "F402", "F823" ]
select = ["C", "E", "F", "I", "W"]

# Ignore import violations in all `__init__.py` files.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[tool.ruff.lint.isort]
lines-after-imports = 2
known-first-party = ["clin34"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

